+++ AltCover.CSApi/Definitions.cs
@@ -13,6 +13,7 @@ namespace AltCover.Parameters
         string Cobertura { get; }
         string OutputFile { get; }
         bool ExposeReturnCode { get; }
+        string SummaryFormat { get; }
         string[] CommandLine { get; }
 
         FSApi.CollectParams ToParameters();
@@ -73,6 +74,11 @@ namespace AltCover.Parameters
     {
         bool FailFast { get; }
     }
+
+    public interface ICLIArg3 : ICLIArg2
+    {
+        string ShowSummary { get; }
+    }
 }
 
 namespace AltCover.Parameters.Primitive
@@ -86,6 +92,7 @@ namespace AltCover.Parameters.Primitive
         public string Threshold { get; set; }
         public string Cobertura { get; set; }
         public string OutputFile { get; set; }
+        public string SummaryFormat { get; set; }
         public string[] CommandLine { get; set; }
         public bool ExposeReturnCode { get; set; }
 
@@ -100,7 +107,8 @@ namespace AltCover.Parameters.Primitive
                 Cobertura,
                 OutputFile,
                 CommandLine,
-                ExposeReturnCode
+                ExposeReturnCode,
+                SummaryFormat
                 );
             return FSApi.CollectParams.NewPrimitive(primitive);
         }
@@ -117,7 +125,8 @@ namespace AltCover.Parameters.Primitive
                 Cobertura = string.Empty,
                 OutputFile = string.Empty,
                 CommandLine = new string[] { },
-                ExposeReturnCode = true
+                ExposeReturnCode = true,
+                SummaryFormat = string.Empty
             };
         }
 
@@ -256,10 +265,11 @@ namespace AltCover.Parameters.Primitive
         }
     }
 
-    public class CLIArgs : ICLIArg2
+    public class CLIArgs : ICLIArg3
     {
         public bool Force { get; set; }
         public bool FailFast { get; set; }
+        public string ShowSummary { get; set; }
     }
 }
 
@@ -294,6 +304,11 @@ namespace AltCover
             var force = DotNet.CLIArgs.NewForce(args.Force);
             switch (args)
             {
+                case ICLIArg3 args3:
+                    var failfast3 = DotNet.CLIArgs.NewFailFast(args3.FailFast);
+                    var showsummary = DotNet.CLIArgs.NewShowSummary(args3.ShowSummary);
+                    return DotNet.CLIArgs.NewMany(new[] { force, failfast3, showsummary });
+
                 case ICLIArg2 args2:
                     var failfast = DotNet.CLIArgs.NewFailFast(args2.FailFast);
                     return DotNet.CLIArgs.NewMany(new[] { force, failfast });